Decompose cli/src/main.rs into focused modules (issue #64)
Split the 2985-line main.rs into client.rs, render.rs, and
commands/{agent,issue,session,spec,worktree,server}.rs, and
added crates/cli/cli-architecture.spec.md.
