+++ crates/cli/src/commands/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod agent;
+pub(crate) mod issue;
+pub(crate) mod server;
+pub(crate) mod session;
+pub(crate) mod spec;
+pub(crate) mod worktree;
