+++ crates/cli/src/main.rs
@@ -1,8 +1,8 @@
 use clap::{Parser, Subcommand};
-use serde_json::json;
-use std::path::PathBuf;
-use types::{Issue, Session};
-use uuid::Uuid;
+
+mod client;
+mod commands;
+mod render;
 
 #[derive(Parser)]
 #[command(name = "ns2")]
@@ -288,442 +288,6 @@ fn load_dotenv() {
     }
 }
 
-pub fn data_dir_and_pid(port: u16) -> (PathBuf, PathBuf) {
-    let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".into());
-    let repo_name = workspace::git_root_sync()
-        .and_then(|p| p.file_name().map(|n| n.to_string_lossy().to_string()))
-        .unwrap_or_else(|| "default".to_string());
-
-    let data_dir = PathBuf::from(&home).join(".ns2").join(&repo_name);
-    let pid_file = data_dir.join(format!("server-{port}.pid"));
-    (data_dir, pid_file)
-}
-
-fn format_issue_row(issue: &Issue) -> String {
-    format!(
-        "{:<6}  {:<30}  {:<10}  {:<12}  {:<25}  {}",
-        issue.id,
-        if issue.title.chars().count() > 30 {
-            issue.title.chars().take(30).collect::<String>()
-        } else {
-            issue.title.clone()
-        },
-        issue.status.to_string(),
-        issue.assignee.as_deref().unwrap_or("-"),
-        if issue.branch.chars().count() > 25 {
-            issue.branch.chars().take(25).collect::<String>()
-        } else {
-            issue.branch.clone()
-        },
-        issue.created_at.format("%Y-%m-%d %H:%M:%S UTC"),
-    )
-}
-
-fn print_issue_row(issue: &Issue) {
-    println!("{}", format_issue_row(issue));
-}
-
-fn handle_connection_error(err: &reqwest::Error) -> ! {
-    if err.is_connect() {
-        eprintln!("Error: server is not running (connection refused). Start it with: ns2 server start");
-    } else {
-        eprintln!("Error: {err}");
-    }
-    std::process::exit(1);
-}
-
-async fn print_error_response(resp: reqwest::Response) -> ! {
-    let status = resp.status();
-    if let Ok(body) = resp.json::<serde_json::Value>().await {
-        if let Some(msg) = body.get("error").and_then(|v| v.as_str()) {
-            eprintln!("Error: {msg}");
-        } else {
-            eprintln!("Error: {status}");
-        }
-    } else {
-        eprintln!("Error: {status}");
-    }
-    std::process::exit(1);
-}
-
-pub fn format_session_event(event: &types::SessionEvent) -> Option<String> {
-    use types::SessionEvent::*;
-    match event {
-        TurnStarted { turn } => Some(format!("[turn {}]\n", turn.id)),
-        ContentBlockDelta {
-            delta: types::ContentBlockDelta::TextDelta { text },
-            ..
-        } => Some(text.clone()),
-        ContentBlockDelta {
-            delta: types::ContentBlockDelta::InputJsonDelta { .. },
-            ..
-        } => None,
-        ContentBlockDone { block, .. } => match block {
-            types::ContentBlock::Text { .. } => Some("\n".to_string()),
-            types::ContentBlock::ToolUse { name, input, .. } => {
-                Some(format!("[tool: {}({})]\n", name, input))
-            }
-            types::ContentBlock::ToolResult { content, .. } => {
-                Some(format!("[result: {}]\n", content))
-            }
-        },
-        TurnDone { .. } => None,
-        SessionDone { .. } => Some("[done]\n".to_string()),
-        Error { message } => Some(format!("[error] {message}\n")),
-    }
-}
-
-fn print_session_event(event: &types::SessionEvent, to_stderr: bool) {
-    use std::io::Write;
-    use types::SessionEvent::*;
-    match event {
-        // Text deltas stream without a newline; flush so the terminal shows them immediately.
-        ContentBlockDelta {
-            delta: types::ContentBlockDelta::TextDelta { .. },
-            ..
-        } => {
-            if let Some(text) = format_session_event(event) {
-                if to_stderr {
-                    eprint!("{text}");
-                    std::io::stderr().flush().ok();
-                } else {
-                    print!("{text}");
-                    std::io::stdout().flush().ok();
-                }
-            }
-        }
-        // Errors always go to stderr.
-        Error { message } => eprintln!("[error] {message}"),
-        // Everything else: print the formatted string (which already includes a newline).
-        _ => {
-            if let Some(output) = format_session_event(event) {
-                if to_stderr {
-                    eprint!("{output}");
-                } else {
-                    print!("{output}");
-                }
-            }
-        }
-    }
-}
-
-pub fn parse_sse_frames(buffer: &mut String, new_data: &str) -> Vec<String> {
-    buffer.push_str(new_data);
-    let mut frames = Vec::new();
-    while let Some(pos) = buffer.find("\n\n") {
-        let frame = buffer[..pos].to_string();
-        *buffer = buffer[pos + 2..].to_string();
-        frames.push(frame);
-    }
-    frames
-}
-
-pub fn format_sync_error(spec_path: &str, stale: &[PathBuf]) -> String {
-    let mut out = format!("[error] spec {spec_path} has stale files:\n");
-    for f in stale {
-        out.push_str(&format!("  {}\n", f.display()));
-    }
-    out
-}
-
-pub fn format_sync_warning(spec_path: &str, stale: &[PathBuf]) -> String {
-    let mut out = format!("[warning] spec {spec_path} has stale files:\n");
-    for f in stale {
-        out.push_str(&format!("  {}\n", f.display()));
-    }
-    out
-}
-
-async fn stream_events(url: &str, to_stderr: bool) {
-    use futures::StreamExt;
-    let client = reqwest::Client::new();
-    let resp = client
-        .get(url)
-        .header("Accept", "text/event-stream")
-        .send()
-        .await
-        .unwrap_or_else(|e| handle_connection_error(&e));
-
-    let mut stream = resp.bytes_stream();
-    let mut buffer = String::new();
-
-    while let Some(chunk) = stream.next().await {
-        let chunk = chunk.unwrap_or_else(|e| {
-            eprintln!("Stream error: {e}");
-            std::process::exit(1);
-        });
-        if let Ok(s) = std::str::from_utf8(&chunk) {
-            let frames = parse_sse_frames(&mut buffer, s);
-            for line in frames {
-                if let Some(data) = line.strip_prefix("data: ") {
-                    if let Ok(event) = serde_json::from_str::<types::SessionEvent>(data) {
-                        print_session_event(&event, to_stderr);
-                        if matches!(event, types::SessionEvent::SessionDone { .. }) {
-                            return;
-                        }
-                        if matches!(event, types::SessionEvent::Error { .. }) {
-                            std::process::exit(1);
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-async fn resolve_session_id(server: &str, id: Option<String>, name: Option<String>) -> Uuid {
-    if let Some(id) = id {
-        id.parse().unwrap_or_else(|_| {
-            eprintln!("Invalid session ID");
-            std::process::exit(1);
-        })
-    } else if let Some(name) = name {
-        let client = reqwest::Client::new();
-        let sessions: Vec<types::Session> = client
-            .get(format!("{server}/sessions"))
-            .send()
-            .await
-            .unwrap()
-            .json()
-            .await
-            .unwrap();
-        sessions
-            .into_iter()
-            .find(|s| s.name == name)
-            .map(|s| s.id)
-            .unwrap_or_else(|| {
-                eprintln!("Session not found: {name}");
-                std::process::exit(1);
-            })
-    } else {
-        eprintln!("Must provide --id or --name");
-        std::process::exit(1);
-    }
-}
-
-pub fn issue_is_terminal(status: &types::IssueStatus) -> bool {
-    matches!(status, types::IssueStatus::Completed | types::IssueStatus::Failed)
-}
-
-/// Check whether every node in the issue tree (roots AND all descendants) is terminal.
-/// Used by `issue wait` to decide when to stop polling.
-pub fn all_nodes_terminal(roots: &[IssueTreeNode]) -> bool {
-    fn node_terminal(node: &IssueTreeNode) -> bool {
-        issue_is_terminal(&node.issue.status) && node.children.iter().all(node_terminal)
-    }
-    roots.iter().all(node_terminal)
-}
-
-/// The result of verifying a batch of spec paths.
-pub struct VerifyResult {
-    /// Lines to print to stdout (one per successfully verified path).
-    pub stdout_lines: Vec<String>,
-    /// Lines to print to stderr (one per failure).
-    pub stderr_lines: Vec<String>,
-    /// Whether any path failed.
-    pub any_failed: bool,
-}
-
-/// Core logic for `ns2 spec verify <paths...>`.
-///
-/// For each path: resolve it relative to `git_root`, attempt to load + write the spec,
-/// record success/failure.  Does NOT call `process::exit` — returns a [`VerifyResult`]
-/// so callers (main and tests) can assert on the outcome.
-pub fn verify_spec_paths(git_root: &std::path::Path, paths: &[String]) -> VerifyResult {
-    let mut stdout_lines = Vec::new();
-    let mut stderr_lines = Vec::new();
-    let mut any_failed = false;
-
-    for path in paths {
-        let resolved = if PathBuf::from(path).is_absolute() {
-            PathBuf::from(path)
-        } else {
-            git_root.join(path)
-        };
-
-        let mut def = match specs::load_spec(&resolved) {
-            Some(d) => d,
-            None => {
-                stderr_lines.push(format!("Error: could not load spec at {path}"));
-                any_failed = true;
-                continue;
-            }
-        };
-
-        def.verified = Some(chrono::Utc::now());
-
-        if let Err(e) = specs::write_spec(&resolved, &def) {
-            stderr_lines.push(format!("Error writing spec file {path}: {e}"));
-            any_failed = true;
-            continue;
-        }
-
-        stdout_lines.push(format!("Verified {path}"));
-    }
-
-    VerifyResult { stdout_lines, stderr_lines, any_failed }
-}
-
-pub fn session_is_terminal(status: &types::SessionStatus) -> bool {
-    matches!(
-        status,
-        types::SessionStatus::Completed | types::SessionStatus::Failed | types::SessionStatus::Cancelled
-    )
-}
-
-/// Braille spinner frames for animated "running" indicator.
-pub const SPINNER_FRAMES: &[char] = &['⠋', '⠙', '⠹', '⠸', '⠼', '⠴', '⠦', '⠧', '⠇', '⠏'];
-
-/// Return the spinner character for the given tick index.
-pub fn spinner_char(tick: usize) -> char {
-    SPINNER_FRAMES[tick % SPINNER_FRAMES.len()]
-}
-
-/// Return the symbol and status label for an issue status.
-pub fn issue_status_symbol(status: &types::IssueStatus, tick: usize) -> (String, &'static str) {
-    match status {
-        types::IssueStatus::Running => (spinner_char(tick).to_string(), "running"),
-        types::IssueStatus::Completed => ("✔".to_string(), "completed"),
-        types::IssueStatus::Failed => ("✗".to_string(), "failed"),
-        types::IssueStatus::Open => ("●".to_string(), "open"),
-    }
-}
-
-/// Truncate a string to at most `max_chars` Unicode characters.
-pub fn truncate_str(s: &str, max_chars: usize) -> String {
-    if s.chars().count() <= max_chars {
-        s.to_string()
-    } else {
-        let truncated: String = s.chars().take(max_chars).collect();
-        format!("{}…", truncated.trim_end())
-    }
-}
-
-/// A node in the issue tree for rendering.
-pub struct IssueTreeNode {
-    pub issue: types::Issue,
-    pub snippet: Option<String>,
-    pub children: Vec<IssueTreeNode>,
-}
-
-/// Render a single tree line for one issue node.
-///
-/// - `prefix` is the indentation/connector string (e.g., "├── ", "└── ", "│   ├── ").
-/// - `tick` is the spinner frame counter.
-/// - `is_root` controls whether to include the snippet.
-pub fn render_tree_line(node: &IssueTreeNode, prefix: &str, tick: usize, is_root: bool) -> String {
-    let (sym, status_label) = issue_status_symbol(&node.issue.status, tick);
-    let id = &node.issue.id;
-
-    let title_part = truncate_str(&node.issue.title, 30);
-
-    if is_root {
-        let snippet_part = if let Some(ref snippet) = node.snippet {
-            // Sanitize: replace newlines/carriage-returns with spaces so the
-            // line-count assumption used for ANSI cursor-up redraw is not broken.
-            let clean: String = snippet.chars().map(|c| if c == '\n' || c == '\r' { ' ' } else { c }).collect();
-            let s = truncate_str(&clean, 30);
-            if s.is_empty() {
-                String::new()
-            } else {
-                format!(": {s}")
-            }
-        } else {
-            String::new()
-        };
-        format!("{prefix}[{id}] {title_part}{snippet_part}  {sym} {status_label}")
-    } else {
-        format!("{prefix}[{id}] {title_part}  {sym} {status_label}")
-    }
-}
-
-/// Render the full issue tree to a vector of lines.
-pub fn render_issue_tree(roots: &[IssueTreeNode], tick: usize) -> Vec<String> {
-    let mut lines = Vec::new();
-    for (ri, root) in roots.iter().enumerate() {
-        let is_last_root = ri + 1 == roots.len();
-        let root_prefix = if roots.len() > 1 {
-            if !is_last_root { "├── " } else { "└── " }
-        } else {
-            ""
-        };
-        lines.push(render_tree_line(root, root_prefix, tick, true));
-        // When there are multiple roots, children must be indented relative to the
-        // root's own connector so the tree lines up correctly:
-        //   ├── Root A          <- root_prefix = "├── "
-        //   │   ├── Child 1     <- parent_indent = "│   "
-        //   │   └── Child 2
-        //   └── Root B          <- root_prefix = "└── "
-        //       └── Child 3     <- parent_indent = "    "
-        let child_indent = if roots.len() > 1 {
-            if !is_last_root { "│   " } else { "    " }
-        } else {
-            ""
-        };
-        render_children(&root.children, child_indent, tick, &mut lines);
-    }
-    lines
-}
-
-fn render_children(children: &[IssueTreeNode], parent_indent: &str, tick: usize, lines: &mut Vec<String>) {
-    for (i, child) in children.iter().enumerate() {
-        let is_last = i + 1 == children.len();
-        let connector = if is_last { "└── " } else { "├── " };
-        let prefix = format!("{parent_indent}{connector}");
-        lines.push(render_tree_line(child, &prefix, tick, false));
-
-        // Recurse into grandchildren
-        let child_indent = if is_last {
-            format!("{parent_indent}    ")
-        } else {
-            format!("{parent_indent}│   ")
-        };
-        render_children(&child.children, &child_indent, tick, lines);
-    }
-}
-
-// ────────────────────────────────────────────────────────────────────────────
-// Session wait progress rendering
-
-/// Return `(symbol, label)` for the given session status at a tick.
-/// Running sessions get an animated braille spinner; terminal sessions get a static symbol.
-pub fn session_status_symbol(status: &types::SessionStatus, tick: usize) -> (String, &'static str) {
-    match status {
-        types::SessionStatus::Running => (spinner_char(tick).to_string(), "running"),
-        types::SessionStatus::Created => (spinner_char(tick).to_string(), "created"),
-        types::SessionStatus::Completed => ("✔".to_string(), "completed"),
-        types::SessionStatus::Failed => ("✗".to_string(), "failed"),
-        types::SessionStatus::Cancelled => ("●".to_string(), "cancelled"),
-    }
-}
-
-/// Render one progress line for a session.
-///
-/// Format: `[<id-prefix>] <name>: <snippet>  <sym> <status>`
-///
-/// - `id` is the full session UUID string; the first 8 chars are shown.
-/// - `name` is the session name (empty string renders as `-`).
-/// - `snippet` is an optional last-content text snippet (truncated to 40 chars).
-pub fn render_session_line(
-    id: &str,
-    name: &str,
-    snippet: Option<&str>,
-    status: &types::SessionStatus,
-    tick: usize,
-) -> String {
-    let id_prefix = &id[..id.len().min(8)];
-    let display_name = if name.is_empty() { "-" } else { name };
-    let (sym, label) = session_status_symbol(status, tick);
-    let snippet_part = match snippet {
-        Some(s) if !s.trim().is_empty() => {
-            let trimmed = s.trim().replace(['\n', '\r'], " ");
-            format!("{}  ", truncate_str(&trimmed, 40))
-        }
-        _ => String::new(),
-    };
-    format!("[{id_prefix}] {display_name}: {snippet_part}{sym} {label}")
-}
-
 #[tokio::main]
 async fn main() {
     load_dotenv();
@@ -732,989 +296,91 @@ async fn main() {
     match cli.command {
         Command::Server { action } => match action {
             ServerAction::Start { port } => {
-                let (data_dir, pid_file) = data_dir_and_pid(port);
-                let config = server::ServerConfig {
-                    port,
-                    data_dir,
-                    pid_file,
-                    model: std::env::var("ANTHROPIC_MODEL")
-                        .unwrap_or_else(|_| "claude-sonnet-4-6".to_string()),
-                };
-                if let Err(e) = server::run(config).await {
-                    eprintln!("Server error: {e}");
-                    std::process::exit(1);
-                }
+                commands::server::run_start(port).await;
             }
             ServerAction::Stop => {
-                let (_, pid_file) = data_dir_and_pid(9876);
-                match std::fs::read_to_string(&pid_file) {
-                    Ok(pid_str) => {
-                        let pid = pid_str.trim().to_string();
-                        // Use sh to invoke kill so the shell builtin is available
-                        // even on minimal systems without a standalone kill binary.
-                        let result = std::process::Command::new("sh")
-                            .args(["-c", &format!("kill -TERM {pid}")])
-                            .output();
-                        match result {
-                            Ok(o) if o.status.success() => {
-                                eprintln!("Server stopped (pid {pid})");
-                            }
-                            Ok(o) => {
-                                let stderr = String::from_utf8_lossy(&o.stderr);
-                                if stderr.contains("No such process") {
-                                    // Stale PID file — process already gone
-                                    let _ = std::fs::remove_file(&pid_file);
-                                    eprintln!("Warning: server process {pid} was not running (stale PID file removed)");
-                                } else {
-                                    eprintln!("Failed to stop server: {stderr}");
-                                    std::process::exit(1);
-                                }
-                            }
-                            Err(e) => {
-                                eprintln!("Failed to send signal: {e}");
-                                std::process::exit(1);
-                            }
-                        }
-                    }
-                    Err(_) => {
-                        eprintln!("No PID file found at {}", pid_file.display());
-                        std::process::exit(1);
-                    }
-                }
+                commands::server::run_stop();
             }
         },
         Command::Session { action } => match action {
             SessionAction::List { status, id } => {
-                let client = reqwest::Client::new();
-
-                // If --id is provided, fetch the specific session
-                if let Some(session_id) = id {
-                    let session_uuid: Uuid = session_id.parse().unwrap_or_else(|_| {
-                        eprintln!("Invalid session ID: {session_id}");
-                        std::process::exit(1);
-                    });
-                    let url = format!("{}/sessions/{}", cli.server, session_uuid);
-                    let resp = client.get(&url).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                            eprintln!("Error: session not found: {session_uuid}");
-                        } else {
-                            eprintln!("Error: {}", resp.status());
-                        }
-                        std::process::exit(1);
-                    }
-                    let session: Session = resp.json().await.unwrap_or_else(|e| {
-                        eprintln!("Error parsing response: {e}");
-                        std::process::exit(1);
-                    });
-                    println!("{:<36}  {:<20}  {:<10}  created_at", "id", "name", "status");
-                    println!(
-                        "{:<36}  {:<20}  {:<10}  {}",
-                        session.id, session.name, session.status, session.created_at
-                    );
-                } else {
-                    // List all sessions with optional status filter
-                    let mut url = format!("{}/sessions", cli.server);
-                    if let Some(s) = &status {
-                        url = format!("{url}?status={s}");
-                    }
-                    let resp = client.get(&url).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        eprintln!("Error: {}", resp.status());
-                        std::process::exit(1);
-                    }
-                    let sessions: Vec<Session> = resp.json().await.unwrap_or_else(|e| {
-                        eprintln!("Error parsing response: {e}");
-                        std::process::exit(1);
-                    });
-                    if sessions.is_empty() {
-                        println!("No sessions found.");
-                    } else {
-                        println!("{:<36}  {:<20}  {:<10}  created_at", "id", "name", "status");
-                        for s in &sessions {
-                            println!(
-                                "{:<36}  {:<20}  {:<10}  {}",
-                                s.id, s.name, s.status, s.created_at
-                            );
-                        }
-                    }
-                }
+                commands::session::run_list(&cli.server, status, id).await;
             }
             SessionAction::New { name, agent, message, wait } => {
-                let url = format!("{}/sessions", cli.server);
-                let body = json!({
-                    "name": name,
-                    "agent": agent,
-                    "initial_message": message,
-                });
-                let client = reqwest::Client::new();
-                let resp = client.post(&url).json(&body).send().await.unwrap_or_else(|e| {
-                    handle_connection_error(&e);
-                });
-                if !resp.status().is_success() {
-                    eprintln!("Error: {}", resp.status());
-                    std::process::exit(1);
-                }
-                let session: Session = resp.json().await.unwrap_or_else(|e| {
-                    eprintln!("Error parsing response: {e}");
-                    std::process::exit(1);
-                });
-                eprintln!("Created session: {} ({})", session.name, session.id);
-                println!("{}", session.id);
-
-                if wait {
-                    // last_turns=1: show only the final turn, not full history.
-                    // Output goes to stderr so stdout stays UUID-only for scripting.
-                    let events_url = format!("{}/sessions/{}/events?last_turns=1", cli.server, session.id);
-                    stream_events(&events_url, true).await;
-                }
+                commands::session::run_new(&cli.server, name, agent, message, wait).await;
             }
             SessionAction::Tail { id, name, turns } => {
-                let session_id = resolve_session_id(&cli.server, id, name).await;
-                let mut url = format!("{}/sessions/{}/events", cli.server, session_id);
-                if let Some(n) = turns {
-                    url = format!("{url}?last_turns={n}");
-                }
-                stream_events(&url, false).await;
+                commands::session::run_tail(&cli.server, id, name, turns).await;
             }
             SessionAction::Send { id, name, message } => {
-                let session_id = resolve_session_id(&cli.server, id, name).await;
-                let url = format!("{}/sessions/{}/messages", cli.server, session_id);
-                let body = serde_json::json!({ "message": message });
-                let client = reqwest::Client::new();
-                let resp = client.post(&url).json(&body).send().await.unwrap_or_else(|e| {
-                    handle_connection_error(&e);
-                });
-                if !resp.status().is_success() {
-                    eprintln!("Error sending message: {}", resp.status());
-                    std::process::exit(1);
-                }
-                eprintln!("Message sent.");
+                commands::session::run_send(&cli.server, id, name, message).await;
             }
             SessionAction::Stop { id, name } => {
-                let session_id = resolve_session_id(&cli.server, id, name).await;
-                // For MVP, just print "not implemented" — real cancellation is out of scope
-                println!("Stop not yet implemented for session {session_id}");
+                commands::session::run_stop(&cli.server, id, name).await;
             }
             SessionAction::Wait { ids } => {
-                if ids.is_empty() {
-                    eprintln!("Error: at least one --id is required");
-                    std::process::exit(1);
-                }
-                let client = reqwest::Client::new();
-                // Validate that all session IDs parse as valid UUIDs up-front.
-                for id in &ids {
-                    if id.parse::<Uuid>().is_err() {
-                        eprintln!("Error: invalid session ID: {id}");
-                        std::process::exit(1);
-                    }
-                }
-
-                use std::io::Write;
-                use std::collections::HashMap;
-
-                let mut terminal_statuses: HashMap<String, types::SessionStatus> = HashMap::new();
-                // Track snippet text per session id for the progress display.
-                let mut snippets: HashMap<String, String> = HashMap::new();
-                // Track session names fetched from the server.
-                let mut names: HashMap<String, String> = HashMap::new();
-                // lines_rendered tracks how many progress lines we've printed so we can
-                // cursor-up to overwrite them on the next tick.
-                let mut lines_rendered: usize = 0;
-                let mut tick: usize = 0;
-
-                loop {
-                    let mut all_done = true;
-                    for id in &ids {
-                        if terminal_statuses.contains_key(id.as_str()) {
-                            continue;
-                        }
-                        let url = format!("{}/sessions/{}", cli.server, id);
-                        let resp = client.get(&url).send().await.unwrap_or_else(|e| {
-                            handle_connection_error(&e);
-                        });
-                        if !resp.status().is_success() {
-                            // Clear progress lines before printing the error.
-                            if lines_rendered > 0 {
-                                eprint!("\x1b[{}A\x1b[J", lines_rendered);
-                            }
-                            if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                                eprintln!("Error: session not found: {id}");
-                            } else {
-                                print_error_response(resp).await;
-                            }
-                            std::process::exit(1);
-                        }
-                        let session: Session = resp.json().await.unwrap_or_else(|e| {
-                            eprintln!("Error parsing response: {e}");
-                            std::process::exit(1);
-                        });
-                        // Cache the name on first fetch.
-                        names.entry(id.clone()).or_insert_with(|| session.name.clone());
-
-                        if session_is_terminal(&session.status) {
-                            terminal_statuses.insert(id.clone(), session.status);
-                        } else {
-                            all_done = false;
-                            // Fetch snippet for running sessions.
-                            let text_url = format!("{}/sessions/{}/last_text", cli.server, id);
-                            if let Ok(text_resp) = client.get(&text_url).send().await {
-                                if text_resp.status().is_success() {
-                                    if let Ok(body) = text_resp.json::<serde_json::Value>().await {
-                                        if let Some(t) = body.get("text").and_then(|v| v.as_str()) {
-                                            snippets.insert(id.clone(), t.to_string());
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    // Render progress lines to stderr.
-                    // On subsequent ticks, move cursor up to overwrite previous lines.
-                    {
-                        let stderr = std::io::stderr();
-                        let mut out = stderr.lock();
-                        if lines_rendered > 0 {
-                            // Move up `lines_rendered` lines and clear to end of screen.
-                            write!(out, "\x1b[{}A\x1b[J", lines_rendered).ok();
-                        }
-                        let mut count = 0;
-                        for id in &ids {
-                            let status = terminal_statuses
-                                .get(id.as_str())
-                                .cloned()
-                                .unwrap_or(types::SessionStatus::Running);
-                            let name = names.get(id.as_str()).map(|s| s.as_str()).unwrap_or("");
-                            let snippet = snippets.get(id.as_str()).map(|s| s.as_str());
-                            let line = render_session_line(id, name, snippet, &status, tick);
-                            writeln!(out, "{line}").ok();
-                            count += 1;
-                        }
-                        lines_rendered = count;
-                    }
-
-                    if all_done {
-                        break;
-                    }
-                    tick = tick.wrapping_add(1);
-                    tokio::time::sleep(tokio::time::Duration::from_millis(250)).await;
-                }
-
-                // Print final static lines to stdout: <uuid>  <status>
-                // (The progress lines on stderr already show the final state.)
-                let mut any_failed = false;
-                for id in &ids {
-                    let status = terminal_statuses.get(id.as_str()).expect("all ids terminal");
-                    println!("{id}  {status}");
-                    if *status == types::SessionStatus::Failed {
-                        any_failed = true;
-                    }
-                }
-                if any_failed {
-                    std::process::exit(1);
-                }
+                commands::session::run_wait(&cli.server, ids).await;
             }
         },
         Command::Agent { action } => match action {
             AgentAction::List => {
-                let dir = agents::agents_dir().unwrap_or_else(|| {
-                    eprintln!("Error: not inside a git repository");
-                    std::process::exit(1);
-                });
-                if !dir.exists() {
-                    println!("No agents found (directory does not exist: {})", dir.display());
-                    return;
-                }
-                let agent_list = agents::list_agents(&dir);
-                if agent_list.is_empty() {
-                    println!("No agents found.");
-                } else {
-                    println!("{:<20} description", "name");
-                    for a in &agent_list {
-                        println!("{:<20} {}", a.name, a.description);
-                    }
-                }
+                commands::agent::run_list();
             }
             AgentAction::New { name, description, body } => {
-                let name = name.unwrap_or_else(|| {
-                    eprintln!("Error: --name is required");
-                    std::process::exit(1);
-                });
-                let dir = agents::agents_dir().unwrap_or_else(|| {
-                    eprintln!("Error: not inside a git repository");
-                    std::process::exit(1);
-                });
-                if let Err(e) = std::fs::create_dir_all(&dir) {
-                    eprintln!("Error creating agents directory: {e}");
-                    std::process::exit(1);
-                }
-                let path = dir.join(format!("{name}.md"));
-                if path.exists() {
-                    eprintln!("Error: agent '{name}' already exists at {}", path.display());
-                    std::process::exit(1);
-                }
-                let open_editor = body.is_none();
-                let def = agents::AgentDef {
-                    name: name.clone(),
-                    description: description.unwrap_or_default(),
-                    body: body.unwrap_or_default(),
-                    include_project_config: false,
-                    hooks: agents::AgentHooks::default(),
-                };
-                if let Err(e) = agents::write_agent(&dir, &def) {
-                    eprintln!("Error writing agent file: {e}");
-                    std::process::exit(1);
-                }
-                if open_editor {
-                    let editor = std::env::var("EDITOR").unwrap_or_else(|_| "vi".to_string());
-                    std::process::Command::new(&editor).arg(&path).status().ok();
-                }
-                eprintln!("Created agent '{name}' at {}", path.display());
+                commands::agent::run_new(name, description, body);
             }
             AgentAction::Edit { name, description, body } => {
-                let name = name.unwrap_or_else(|| {
-                    eprintln!("Error: --name is required");
-                    std::process::exit(1);
-                });
-                if description.is_none() && body.is_none() {
-                    eprintln!("Error: at least one of --description or --body must be provided");
-                    std::process::exit(1);
-                }
-                let dir = agents::agents_dir().unwrap_or_else(|| {
-                    eprintln!("Error: not inside a git repository");
-                    std::process::exit(1);
-                });
-                let mut def = agents::load_agent(&dir, &name).unwrap_or_else(|| {
-                    eprintln!(
-                        "Error: agent '{name}' not found at {}",
-                        dir.join(format!("{name}.md")).display()
-                    );
-                    std::process::exit(1);
-                });
-                if let Some(d) = description {
-                    def.description = d;
-                }
-                if let Some(b) = body {
-                    def.body = b;
-                }
-                if let Err(e) = agents::write_agent(&dir, &def) {
-                    eprintln!("Error writing agent file: {e}");
-                    std::process::exit(1);
-                }
-                eprintln!("Updated agent '{name}'.");
+                commands::agent::run_edit(name, description, body);
             }
         },
         Command::Spec { action } => match action {
             SpecAction::New { path, targets, severity } => {
-                if targets.is_empty() {
-                    eprintln!("Error: at least one --target is required");
-                    std::process::exit(1);
-                }
-                let severity = match severity.as_str() {
-                    "warning" => specs::Severity::Warning,
-                    "error" => specs::Severity::Error,
-                    _ => {
-                        eprintln!("Error: --severity must be 'error' or 'warning'");
-                        std::process::exit(1);
-                    }
-                };
-                let git_root = workspace::git_root_sync().unwrap_or_else(|| {
-                    eprintln!("Error: not inside a git repository");
-                    std::process::exit(1);
-                });
-                let resolved = if PathBuf::from(&path).is_absolute() {
-                    PathBuf::from(&path)
-                } else {
-                    git_root.join(&path)
-                };
-                let path_display = path.clone();
-                let path = resolved;
-                if path.exists() {
-                    eprintln!("Error: spec already exists at {}", path.display());
-                    std::process::exit(1);
-                }
-                if let Some(parent) = path.parent() {
-                    if !parent.as_os_str().is_empty() {
-                        if let Err(e) = std::fs::create_dir_all(parent) {
-                            eprintln!("Error creating directories: {e}");
-                            std::process::exit(1);
-                        }
-                    }
-                }
-                let def = specs::SpecDef { targets, verified: None, severity, body: String::new() };
-                if let Err(e) = specs::write_spec(&path, &def) {
-                    eprintln!("Error writing spec file: {e}");
-                    std::process::exit(1);
-                }
-                println!("Created spec at {path_display}");
+                commands::spec::run_new(path, targets, severity);
             }
             SpecAction::Sync { path, error_on_warnings } => {
-                let git_root = workspace::git_root_sync().unwrap_or_else(|| {
-                    eprintln!("Error: not inside a git repository");
-                    std::process::exit(1);
-                });
-                if let Some(p) = path {
-                    let resolved = if PathBuf::from(&p).is_absolute() {
-                        PathBuf::from(&p)
-                    } else {
-                        git_root.join(&p)
-                    };
-                    if resolved.is_dir() {
-                        let all_specs = specs::list_specs(&resolved);
-                        let mut has_errors = false;
-                        for (spec_path, def) in &all_specs {
-                            let stale = specs::stale_files(&git_root, spec_path, def);
-                            if !stale.is_empty() {
-                                let display_path = spec_path
-                                    .strip_prefix(&git_root)
-                                    .unwrap_or(spec_path)
-                                    .display()
-                                    .to_string();
-                                let is_error =
-                                    def.severity == specs::Severity::Error || error_on_warnings;
-                                if is_error {
-                                    eprint!("{}", format_sync_error(&display_path, &stale));
-                                    has_errors = true;
-                                } else {
-                                    eprint!("{}", format_sync_warning(&display_path, &stale));
-                                }
-                            }
-                        }
-                        if has_errors {
-                            std::process::exit(1);
-                        }
-                    } else {
-                        let def = specs::load_spec(&resolved).unwrap_or_else(|| {
-                            eprintln!("Error: could not load spec at {p}");
-                            std::process::exit(1);
-                        });
-                        let stale = specs::stale_files(&git_root, &resolved, &def);
-                        if !stale.is_empty() {
-                            let is_error =
-                                def.severity == specs::Severity::Error || error_on_warnings;
-                            if is_error {
-                                eprint!("{}", format_sync_error(&p, &stale));
-                                std::process::exit(1);
-                            } else {
-                                eprint!("{}", format_sync_warning(&p, &stale));
-                            }
-                        }
-                    }
-                } else {
-                    let all_specs = specs::list_specs(&git_root);
-                    let mut has_errors = false;
-                    for (spec_path, def) in &all_specs {
-                        let stale = specs::stale_files(&git_root, spec_path, def);
-                        if !stale.is_empty() {
-                            let display_path = spec_path
-                                .strip_prefix(&git_root)
-                                .unwrap_or(spec_path)
-                                .display()
-                                .to_string();
-                            let is_error =
-                                def.severity == specs::Severity::Error || error_on_warnings;
-                            if is_error {
-                                eprint!("{}", format_sync_error(&display_path, &stale));
-                                has_errors = true;
-                            } else {
-                                eprint!("{}", format_sync_warning(&display_path, &stale));
-                            }
-                        }
-                    }
-                    if has_errors {
-                        std::process::exit(1);
-                    }
-                }
+                commands::spec::run_sync(path, error_on_warnings);
             }
             SpecAction::Verify { paths } => {
-                let git_root = workspace::git_root_sync().unwrap_or_else(|| {
-                    eprintln!("Error: not inside a git repository");
-                    std::process::exit(1);
-                });
-                let result = verify_spec_paths(&git_root, &paths);
-                for line in &result.stdout_lines {
-                    println!("{line}");
-                }
-                for line in &result.stderr_lines {
-                    eprintln!("{line}");
-                }
-                if result.any_failed {
-                    std::process::exit(1);
-                }
+                commands::spec::run_verify(paths);
             }
         },
-        Command::Issue { action } => {
-            let client = reqwest::Client::new();
-            match action {
-                IssueAction::New { title, body, assignee, parent, blocked_on, start, branch } => {
-                    if start && assignee.is_none() {
-                        eprintln!("Error: --start requires --assignee (start needs an agent to run)");
-                        std::process::exit(1);
-                    }
-                    if let Some(ref a) = assignee {
-                        if let Some(dir) = agents::agents_dir() {
-                            if agents::load_agent(&dir, a).is_none() {
-                                eprintln!("Error: agent type '{a}' not found in .ns2/agents/");
-                                std::process::exit(1);
-                            }
-                        }
-                    }
-                    let url = format!("{}/issues", cli.server);
-                    let req_body = json!({
-                        "title": title,
-                        "body": body,
-                        "assignee": assignee,
-                        "parent_id": parent,
-                        "blocked_on": blocked_on,
-                        "branch": branch,
-                    });
-                    let resp = client.post(&url).json(&req_body).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        print_error_response(resp).await;
-                    }
-                    let issue: Issue = resp.json().await.unwrap_or_else(|e| {
-                        eprintln!("Error parsing response: {e}");
-                        std::process::exit(1);
-                    });
-                    eprintln!("Created issue: {} ({})", issue.title, issue.id);
-                    println!("{}", issue.id);
-
-                    if start {
-                        let start_url = format!("{}/issues/{}/start", cli.server, issue.id);
-                        let start_resp = client.post(&start_url).send().await.unwrap_or_else(|e| {
-                            handle_connection_error(&e);
-                        });
-                        if !start_resp.status().is_success() {
-                            if start_resp.status() == reqwest::StatusCode::NOT_FOUND {
-                                eprintln!("Error: issue not found: {}", issue.id);
-                                std::process::exit(1);
-                            }
-                            print_error_response(start_resp).await;
-                        }
-                        let started: Issue = start_resp.json().await.unwrap_or_else(|e| {
-                            eprintln!("Error parsing response: {e}");
-                            std::process::exit(1);
-                        });
-                        eprintln!("Started issue {}. Session: {}", started.id, started.session_id.map(|id| id.to_string()).unwrap_or_default());
-                    }
-                }
-                IssueAction::Edit { id, title, body, assignee, parent, blocked_on, branch } => {
-                    let url = format!("{}/issues/{}", cli.server, id);
-                    let mut req_body = serde_json::Map::new();
-                    if let Some(t) = title {
-                        req_body.insert("title".into(), json!(t));
-                    }
-                    if let Some(b) = body {
-                        req_body.insert("body".into(), json!(b));
-                    }
-                    if let Some(a) = assignee {
-                        if a.is_empty() {
-                            req_body.insert("assignee".into(), json!(null));
-                        } else {
-                            if let Some(dir) = agents::agents_dir() {
-                                if agents::load_agent(&dir, &a).is_none() {
-                                    eprintln!("Error: agent type '{a}' not found in .ns2/agents/");
-                                    std::process::exit(1);
-                                }
-                            }
-                            req_body.insert("assignee".into(), json!(a));
-                        }
-                    }
-                    if let Some(p) = parent {
-                        if p.is_empty() {
-                            req_body.insert("parent_id".into(), json!(null));
-                        } else {
-                            req_body.insert("parent_id".into(), json!(p));
-                        }
-                    }
-                    if let Some(bo) = blocked_on {
-                        req_body.insert("blocked_on".into(), json!(bo));
-                    }
-                    if let Some(br) = branch {
-                        req_body.insert("branch".into(), json!(br));
-                    }
-                    let resp = client
-                        .patch(&url)
-                        .json(&serde_json::Value::Object(req_body))
-                        .send()
-                        .await
-                        .unwrap_or_else(|e| handle_connection_error(&e));
-                    if !resp.status().is_success() {
-                        if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                            eprintln!("Error: issue not found: {id}");
-                            std::process::exit(1);
-                        }
-                        print_error_response(resp).await;
-                    }
-                    let issue: Issue = resp.json().await.unwrap_or_else(|e| {
-                        eprintln!("Error parsing response: {e}");
-                        std::process::exit(1);
-                    });
-                    eprintln!("Updated issue {}.", issue.id);
-                }
-                IssueAction::Comment { id, body, author } => {
-                    let url = format!("{}/issues/{}/comments", cli.server, id);
-                    let req_body = json!({ "author": author, "body": body });
-                    let resp = client.post(&url).json(&req_body).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                            eprintln!("Error: issue not found: {id}");
-                            std::process::exit(1);
-                        }
-                        print_error_response(resp).await;
-                    }
-                    eprintln!("Comment added to issue {id}.");
-                }
-                IssueAction::Start { id } => {
-                    let url = format!("{}/issues/{}/start", cli.server, id);
-                    let resp = client.post(&url).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                            eprintln!("Error: issue not found: {id}");
-                            std::process::exit(1);
-                        }
-                        print_error_response(resp).await;
-                    }
-                    let issue: Issue = resp.json().await.unwrap_or_else(|e| {
-                        eprintln!("Error parsing response: {e}");
-                        std::process::exit(1);
-                    });
-                    eprintln!("Started issue {id}. Session: {}", issue.session_id.map(|id| id.to_string()).unwrap_or_default());
-                }
-                IssueAction::Complete { id, comment } => {
-                    let url = format!("{}/issues/{}/complete", cli.server, id);
-                    let req_body = json!({ "comment": comment });
-                    let resp = client.post(&url).json(&req_body).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                            eprintln!("Error: issue not found: {id}");
-                            std::process::exit(1);
-                        }
-                        print_error_response(resp).await;
-                    }
-                    eprintln!("Issue {id} marked as completed.");
-                }
-                IssueAction::Reopen { id, comment, start } => {
-                    let url = format!("{}/issues/{}/reopen", cli.server, id);
-                    let req_body = json!({ "comment": comment });
-                    let resp = client.post(&url).json(&req_body).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if resp.status() == reqwest::StatusCode::NOT_FOUND {
-                        eprintln!("Error: issue not found: {id}");
-                        std::process::exit(1);
-                    }
-                    if !resp.status().is_success() {
-                        print_error_response(resp).await;
-                    }
-                    eprintln!("Issue {id} reopened.");
-
-                    if start {
-                        let start_url = format!("{}/issues/{}/start", cli.server, id);
-                        let start_resp = client.post(&start_url).send().await.unwrap_or_else(|e| {
-                            handle_connection_error(&e);
-                        });
-                        if !start_resp.status().is_success() {
-                            if start_resp.status() == reqwest::StatusCode::NOT_FOUND {
-                                eprintln!("Error: issue not found: {id}");
-                                std::process::exit(1);
-                            }
-                            print_error_response(start_resp).await;
-                        }
-                        let started: Issue = start_resp.json().await.unwrap_or_else(|e| {
-                            eprintln!("Error parsing response: {e}");
-                            std::process::exit(1);
-                        });
-                        eprintln!("Started issue {id}. Session: {}", started.session_id.map(|id| id.to_string()).unwrap_or_default());
-                    }
-                }
-                IssueAction::List { status, assignee, parent, blocked_on } => {
-                    let mut url = format!("{}/issues", cli.server);
-                    let mut params: Vec<String> = vec![];
-                    if let Some(s) = &status {
-                        params.push(format!("status={s}"));
-                    }
-                    if let Some(a) = &assignee {
-                        params.push(format!("assignee={a}"));
-                    }
-                    if let Some(p) = &parent {
-                        params.push(format!("parent_id={p}"));
-                    }
-                    if let Some(bo) = &blocked_on {
-                        params.push(format!("blocked_on={bo}"));
-                    }
-                    if !params.is_empty() {
-                        url = format!("{url}?{}", params.join("&"));
-                    }
-                    let resp = client.get(&url).send().await.unwrap_or_else(|e| {
-                        handle_connection_error(&e);
-                    });
-                    if !resp.status().is_success() {
-                        eprintln!("Error: {}", resp.status());
-                        std::process::exit(1);
-                    }
-                    let issues: Vec<Issue> = resp.json().await.unwrap_or_else(|e| {
-                        eprintln!("Error parsing response: {e}");
-                        std::process::exit(1);
-                    });
-                    if issues.is_empty() {
-                        println!("No issues found.");
-                    } else {
-                        println!("{:<6}  {:<30}  {:<10}  {:<12}  {:<25}  created_at", "id", "title", "status", "assignee", "branch");
-                        for issue in &issues {
-                            print_issue_row(issue);
-                        }
-                    }
-                }
-                IssueAction::Wait { ids } => {
-                    if ids.is_empty() {
-                        eprintln!("Error: at least one --id is required");
-                        std::process::exit(1);
-                    }
-
-                    // Helper: fetch an issue tree rooted at `id` recursively.
-                    async fn fetch_issue_tree(
-                        client: &reqwest::Client,
-                        server: &str,
-                        id: &str,
-                    ) -> Option<IssueTreeNode> {
-                        let url = format!("{}/issues/{}", server, id);
-                        let resp = client.get(&url).send().await.ok()?;
-                        if !resp.status().is_success() {
-                            return None;
-                        }
-                        let issue: types::Issue = resp.json().await.ok()?;
-
-                        // Fetch children (issues with this parent_id)
-                        let children_url = format!("{}/issues?parent_id={}", server, id);
-                        let children_resp = client.get(&children_url).send().await.ok()?;
-                        let child_issues: Vec<types::Issue> = if children_resp.status().is_success() {
-                            children_resp.json().await.unwrap_or_default()
-                        } else {
-                            vec![]
-                        };
-
-                        // Recursively fetch children
-                        let mut children = Vec::new();
-                        for child in &child_issues {
-                            // Use Box::pin to handle recursive async
-                            if let Some(node) = Box::pin(fetch_issue_tree(client, server, &child.id)).await {
-                                children.push(node);
-                            }
-                        }
-
-                        Some(IssueTreeNode { issue, snippet: None, children })
-                    }
-
-                    // Helper: fetch last text snippet for a running issue with a session.
-                    async fn fetch_snippet(
-                        client: &reqwest::Client,
-                        server: &str,
-                        session_id: uuid::Uuid,
-                    ) -> Option<String> {
-                        let url = format!("{}/sessions/{}/last_text", server, session_id);
-                        let resp = client.get(&url).send().await.ok()?;
-                        if !resp.status().is_success() {
-                            return None;
-                        }
-                        let v: serde_json::Value = resp.json().await.ok()?;
-                        v["text"].as_str().map(|s| s.to_string())
-                    }
-
-                    // Recursively attach snippets to running nodes.
-                    fn attach_snippets(
-                        node: &mut IssueTreeNode,
-                        snippets: &std::collections::HashMap<uuid::Uuid, Option<String>>,
-                    ) {
-                        if node.issue.status == types::IssueStatus::Running {
-                            if let Some(session_id) = node.issue.session_id {
-                                if let Some(snippet_opt) = snippets.get(&session_id) {
-                                    node.snippet = snippet_opt.clone();
-                                }
-                            }
-                        }
-                        for child in &mut node.children {
-                            attach_snippets(child, snippets);
-                        }
-                    }
-
-                    // Collect all running session IDs from a tree.
-                    fn collect_running_sessions(node: &IssueTreeNode, out: &mut Vec<uuid::Uuid>) {
-                        if node.issue.status == types::IssueStatus::Running {
-                            if let Some(sid) = node.issue.session_id {
-                                out.push(sid);
-                            }
-                        }
-                        for child in &node.children {
-                            collect_running_sessions(child, out);
-                        }
-                    }
-
-                    use std::io::Write;
-
-                    let mut tick: usize = 0;
-                    let mut prev_line_count = 0usize;
-                    let mut any_failed = false;
-                    let mut final_statuses: Vec<(String, types::IssueStatus)> = Vec::new();
-
-                    loop {
-                        // Fetch trees for all root IDs
-                        let mut roots: Vec<IssueTreeNode> = Vec::new();
-                        let mut fetch_error = false;
-                        for id in &ids {
-                            match fetch_issue_tree(&client, &cli.server, id).await {
-                                Some(node) => roots.push(node),
-                                None => {
-                                    eprintln!("Error: issue not found: {id}");
-                                    fetch_error = true;
-                                }
-                            }
-                        }
-                        if fetch_error {
-                            std::process::exit(1);
-                        }
-
-                        // Collect running sessions and fetch snippets
-                        let mut session_ids: Vec<uuid::Uuid> = Vec::new();
-                        for root in &roots {
-                            collect_running_sessions(root, &mut session_ids);
-                        }
-                        let mut snippets: std::collections::HashMap<uuid::Uuid, Option<String>> =
-                            std::collections::HashMap::new();
-                        for sid in &session_ids {
-                            let snippet = fetch_snippet(&client, &cli.server, *sid).await;
-                            snippets.insert(*sid, snippet);
-                        }
-
-                        // Attach snippets to running nodes
-                        for root in &mut roots {
-                            attach_snippets(root, &snippets);
-                        }
-
-                        // Render the tree
-                        let lines = render_issue_tree(&roots, tick);
-
-                        // Clear previous frame (cursor-up + clear line for each previous line)
-                        let stderr = std::io::stderr();
-                        let mut out = stderr.lock();
-                        for _ in 0..prev_line_count {
-                            write!(out, "\x1b[1A\x1b[2K").ok();
-                        }
-
-                        // Write new frame
-                        for line in &lines {
-                            writeln!(out, "{}", line).ok();
-                        }
-                        out.flush().ok();
-                        prev_line_count = lines.len();
-
-                        // Check if all done
-                        if all_nodes_terminal(&roots) {
-                            for root in &roots {
-                                if root.issue.status == types::IssueStatus::Failed {
-                                    any_failed = true;
-                                }
-                                final_statuses.push((root.issue.id.clone(), root.issue.status.clone()));
-                            }
-                            break;
-                        }
-
-                        tick = tick.wrapping_add(1);
-                        tokio::time::sleep(tokio::time::Duration::from_millis(250)).await;
-                    }
-
-                    // Print stdout contract: one line per waited issue
-                    for (id, status) in &final_statuses {
-                        println!("{id}  {status}");
-                    }
-
-                    if any_failed {
-                        eprintln!("One or more issues failed.");
-                        std::process::exit(1);
-                    }
-                }
+        Command::Issue { action } => match action {
+            IssueAction::New { title, body, assignee, parent, blocked_on, start, branch } => {
+                commands::issue::run_new(&cli.server, title, body, assignee, parent, blocked_on, start, branch).await;
             }
-        }
-        Command::Worktree { action } => {
-            let git_root = workspace::git_root_sync().unwrap_or_else(|| {
-                eprintln!("Error: not inside a git repository");
-                std::process::exit(1);
-            });
-            let config = workspace::read_ns2_config(&git_root);
-
-            match action {
-                WorktreeAction::List => {
-                    let entries = workspace::list_worktrees(&git_root, &config.worktree_base).await;
-                    if entries.is_empty() {
-                        println!("No worktrees found.");
-                    } else {
-                        println!("{:<40}  path", "branch");
-                        for entry in &entries {
-                            println!(
-                                "{:<40}  {}",
-                                entry.branch,
-                                entry.path.display()
-                            );
-                        }
-                    }
-                }
-                WorktreeAction::Create { branch } => {
-                    let worktree_path = config.worktree_base.join(&branch);
-                    match workspace::ensure_worktree(&git_root, &worktree_path, &branch).await {
-                        Some(path) => {
-                            eprintln!(
-                                "Created worktree for branch {} at {}",
-                                branch,
-                                path.display()
-                            );
-                        }
-                        None => {
-                            eprintln!("Error: failed to create worktree for branch {branch}");
-                            std::process::exit(1);
-                        }
-                    }
-                }
-                WorktreeAction::Delete { branch, force } => {
-                    match workspace::delete_worktree(
-                        &git_root,
-                        &config.worktree_base,
-                        &branch,
-                        force,
-                    ).await {
-                        Ok(_path) => {
-                            eprintln!("Deleted worktree for branch {branch}");
-                        }
-                        Err(workspace::DeleteWorktreeError::NotFound(_)) => {
-                            eprintln!("Error: no worktree found for branch {branch}");
-                            std::process::exit(1);
-                        }
-                        Err(workspace::DeleteWorktreeError::UnmergedCommits(_)) => {
-                            eprintln!(
-                                "Error: branch {branch} has unmerged commits. Use --force to delete anyway."
-                            );
-                            std::process::exit(1);
-                        }
-                        Err(workspace::DeleteWorktreeError::GitFailed(msg)) => {
-                            eprintln!("Error: {msg}");
-                            std::process::exit(1);
-                        }
-                    }
-                }
+            IssueAction::Edit { id, title, body, assignee, parent, blocked_on, branch } => {
+                commands::issue::run_edit(&cli.server, id, title, body, assignee, parent, blocked_on, branch).await;
             }
-        }
+            IssueAction::Comment { id, body, author } => {
+                commands::issue::run_comment(&cli.server, id, body, author).await;
+            }
+            IssueAction::Start { id } => {
+                commands::issue::run_start(&cli.server, id).await;
+            }
+            IssueAction::Complete { id, comment } => {
+                commands::issue::run_complete(&cli.server, id, comment).await;
+            }
+            IssueAction::Reopen { id, comment, start } => {
+                commands::issue::run_reopen(&cli.server, id, comment, start).await;
+            }
+            IssueAction::List { status, assignee, parent, blocked_on } => {
+                commands::issue::run_list(&cli.server, status, assignee, parent, blocked_on).await;
+            }
+            IssueAction::Wait { ids } => {
+                commands::issue::run_wait(&cli.server, ids).await;
+            }
+        },
+        Command::Worktree { action } => match action {
+            WorktreeAction::List => {
+                commands::worktree::run_list().await;
+            }
+            WorktreeAction::Create { branch } => {
+                commands::worktree::run_create(branch).await;
+            }
+            WorktreeAction::Delete { branch, force } => {
+                commands::worktree::run_delete(branch, force).await;
+            }
+        },
     }
 }
 
@@ -1723,6 +389,16 @@ mod tests {
     use super::*;
     use types::*;
     use uuid::Uuid;
+    use crate::render::{
+        format_issue_row, format_session_event, format_sync_error, format_sync_warning,
+        issue_status_symbol, parse_sse_frames, render_issue_tree, render_session_line,
+        render_tree_line, session_status_symbol, spinner_char, truncate_str, IssueTreeNode,
+        SPINNER_FRAMES,
+    };
+    use crate::commands::spec::verify_spec_paths;
+    use crate::commands::issue::{issue_is_terminal, all_nodes_terminal};
+    use crate::commands::session::session_is_terminal;
+    use crate::commands::server::data_dir_and_pid;
 
     #[test]
     fn issue_is_terminal_completed_is_true() {
